[tmux.conf] Don't allow + zoom-in on single-pane window
This was an exercise in learning how horrible it is to do conditional
statements in tmux 2.6 (as packaged for Ubuntu 18.04).  I was thrown off
for a while by looking at online documentation and StackOverflow answers
that used the brace syntax for compound statements which is much nicer,
but this syntax is apparently not available in our 4-year-old version of
tmux.